pyproject.toml
[tool.poetry]
name = "repl-nix-applai-streamline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.46.0"
flask = "^3.0.3"
stripe = "^10.12.0"
streamlit = "^1.38.0"
firebase-admin = "^6.5.0"
pdfminer-six = "^20240706"
python-docx = "^1.1.2"
langchain = "^0.3.0"
llama-index = "^0.11.10"
langchain-community = "^0.3.0"
pandas = "^2.2.2"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
